Revert "#1356 Fix integer custom field"
This reverts commit 3506d12c325a11c2fb2a5099403e63303422270d.

+++ thehive/app/org/thp/thehive/models/Case.scala
@@ -41,14 +41,14 @@ case class CaseCustomField(
     order: Option[Int] = None,
     stringValue: Option[String] = None,
     booleanValue: Option[Boolean] = None,
-    integerValue: Option[Long] = None,
+    integerValue: Option[Int] = None,
     floatValue: Option[Double] = None,
     dateValue: Option[Date] = None
 ) extends CustomFieldValue[CaseCustomField] {
   override def order_=(value: Option[Int]): CaseCustomField            = copy(order = value)
   override def stringValue_=(value: Option[String]): CaseCustomField   = copy(stringValue = value)
   override def booleanValue_=(value: Option[Boolean]): CaseCustomField = copy(booleanValue = value)
-  override def integerValue_=(value: Option[Long]): CaseCustomField    = copy(integerValue = value)
+  override def integerValue_=(value: Option[Int]): CaseCustomField     = copy(integerValue = value)
   override def floatValue_=(value: Option[Double]): CaseCustomField    = copy(floatValue = value)
   override def dateValue_=(value: Option[Date]): CaseCustomField       = copy(dateValue = value)
 }
